(后端)一些完善;

+++ back/postcard-maker/src/main/java/com/louie/coding/dao/PostcardContentDao.java
@@ -7,5 +7,5 @@ import org.apache.ibatis.annotations.Mapper;
 public interface PostcardContentDao {
     void addPostcardContent(PostcardContent postcardContent);
 
-    Integer getMaxVersionByPostcardId(Long postcardId);
+    Long getMaxVersionByPostcardId(Long postcardId);
 }
